docs(List): split ListItem into its own navigation entry
The List documentation page had grown very long. Give List its own
component category (like Chat) and document ListItem on a separate page
next to List, so both are easier to scan.

- New "List" category: List moves from /components/structure/list to
  /components/list/list, ListItem is a new page at /components/list/list-item.
- Move all ListItem content (anatomy, Link, Accordion, Checkboxes, Content
  Slots, ColumnLayout, ActionGroup) to the ListItem page; List keeps a short
  cross-linked teaser.
- Add a Best Practices section and an "Als Raster" (tiles) section to the
  ListItem page; add focused hero and tile examples.
- Remove the "Responsive Layout" section and its mobile example.
- Update internal links pointing to the old List path.

Closes #2877
