+++ apps/docs/src/content/components/list/list-item/examples/item-action-group.tsx
@@ -12,7 +12,7 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list-item/examples/domainApi";
 
 export default () => {
   const List = typedList<Domain>();
