+++ apps/docs/src/content/components/list/list-item/examples/item-checkbox.tsx
@@ -9,7 +9,7 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list-item/examples/domainApi";
 import { useState } from "react";
 
 export default () => {
