+++ apps/docs/src/content/components/list/list-item/examples/item-content.tsx
@@ -1,5 +1,6 @@
 import {
   Avatar,
+  Content,
   ContextMenu,
   Heading,
   IconDomain,
@@ -9,7 +10,7 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list-item/examples/domainApi";
 
 export default () => {
   const List = typedList<Domain>();
@@ -23,7 +24,7 @@ export default () => {
     >
       <List.StaticData data={domains} />
       <List.Item
-        href={() => "#"}
+        showTiles
         textValue={(domain) => domain.domain}
       >
         {(domain) => (
@@ -32,6 +33,9 @@ export default () => {
               <IconDomain />
             </Avatar>
             <Heading>{domain.hostname}</Heading>
+
+            <Content slot="top">Top Content</Content>
+            <Content slot="bottom">Bottom Content</Content>
 
             <ContextMenu>
               <MenuItem>Details anzeigen</MenuItem>
