+++ apps/docs/src/content/components/list/list-item/examples/tile.tsx
@@ -1,41 +1,55 @@
 import {
-  ActionGroup,
   Avatar,
-  Button,
+  ContextMenu,
   Heading,
   IconDomain,
+  IconSubdomain,
+  MenuItem,
   Text,
   typedList,
 } from "@mittwald/flow-react-components";
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list-item/examples/domainApi";
 
 export default () => {
   const DomainList = typedList<Domain>();
 
   return (
     <DomainList.List
-      batchSize={2}
-      hidePagination
+      batchSize={3}
       aria-label="Domains"
+      hidePagination
+      defaultViewMode="tiles"
       getItemId={(domain) => domain.id}
     >
       <DomainList.StaticData data={domains} />
-      <ActionGroup>
-        <Button>Anlegen</Button>
-      </ActionGroup>
       <DomainList.Item
         textValue={(domain) => domain.domain}
+        showTiles
+        showList={false}
       >
         {(domain) => (
           <DomainList.ItemView>
-            <Avatar>
-              <IconDomain />
+            <Avatar
+              color={
+                domain.type === "Domain" ? "blue" : "teal"
+              }
+            >
+              {domain.type === "Domain" ? (
+                <IconDomain />
+              ) : (
+                <IconSubdomain />
+              )}
             </Avatar>
             <Heading>{domain.hostname}</Heading>
             <Text>{domain.type}</Text>
+
+            <ContextMenu>
+              <MenuItem>Details anzeigen</MenuItem>
+              <MenuItem>Löschen</MenuItem>
+            </ContextMenu>
           </DomainList.ItemView>
         )}
       </DomainList.Item>
