+++ apps/docs/src/content/components/list/list-item/index.mdx
@@ -1,0 +1,122 @@
+---
+title: ListItem
+description:
+  Ein ListItem stellt die wichtigsten Informationen eines Elements übersichtlich
+  dar.
+gitHubComponentPath: components/List/components/ListItemView
+---
+
+<LiveCodeEditor />
+
+---
+
+# Best Practices
+
+- Halte den Titel kurz und eindeutig. Lange Headings brechen um und erschweren
+  das schnelle Scannen der List.
+- Zeige nur die Informationen, die zum Verständnis des Elements nötig sind. Zu
+  viele Details machen die List unübersichtlich.
+- Ergänze im Untertitel weitere Informationen nach dem Muster „Beschreibung – 1.
+  Information – 2. Information“.
+- Verlagere selten benötigte Details in Top oder Bottom Content oder blende sie
+  über [ColumnLayout](/components/structure/column-layout) auf kleinen
+  Bildschirmen aus.
+- Gib jedem ListItem über `textValue` seinen Text mit. Erst damit findet der
+  User es über die Tastatur, indem er dessen Anfang tippt.
+
+---
+
+# Aufbau
+
+Ein ListItem wird immer innerhalb einer [List](/components/list/list) gerendert
+und repräsentiert ein spezifisches Element einer Kategorie (z. B. eine Domain,
+E-Mail-Adresse oder ein Projekt). In der Listenansicht besteht es typischerweise
+aus:
+
+- **Avatar:** Der [Avatar](/components/content/avatar) steht am Anfang. Ein
+  [Icon](/components/content/icon) spiegelt die Kategorie wider (z. B. ein
+  Domain-Icon); ein hochgeladenes [Image](/components/content/image) wird
+  angezeigt, andernfalls erscheinen Initialen.
+- **Titel und Untertitel:** Der Titel gibt den Namen wieder, der Untertitel
+  ergänzt weitere Informationen nach dem Muster „Beschreibung – 1. Information
+  – 2. Information“.
+- **Content Slots:** Zusätzliche Bereiche (Top und Bottom Content), die flexibel
+  befüllt werden können.
+- **Aktionen:** Interaktionen erfolgen über ein
+  [ContextMenu](/components/actions/context-menu) oder direkt als
+  [Buttons](/components/actions/button) im ListItem.
+
+---
+
+# Als Raster
+
+In der Rasteransicht wird das ListItem kompakter dargestellt: Der Avatar wird
+größer und eckig, Top Content sowie die Accordion-Funktion entfallen. Aktiviere
+die Darstellung über `showTiles`.
+
+<LiveCodeEditor example="tile" editorCollapsed />
+
+---
+
+# Mit Link
+
+Ein ListItem bietet das Property `href`, um das Element zu verlinken.
+
+<LiveCodeEditor example="item-link" editorCollapsed />
+
+---
+
+# Mit Accordion
+
+Das Accordion-Verhalten wird über die `accordion`-Property aktiviert. Dadurch
+lässt sich ein ListItem per Klick ein- oder ausklappen. Der erweiterte Inhalt
+wird in `<Content slot="bottom" />` platziert.
+
+<LiveCodeEditor example="item-accordion" editorCollapsed />
+
+---
+
+# Mit Checkboxen
+
+[Checkboxen](/components/form-controls/checkbox) in einem ListItem werden
+automatisch am Anfang der Zeile angeordnet. Ihre Funktionalität wird nicht von
+der List gesteuert und muss individuell implementiert werden. Achte darauf, dass
+die gesamte Zeile zur Auswahl genutzt werden kann – nutze dafür `onAction` der
+List.
+
+<LiveCodeEditor example="item-checkbox" editorCollapsed />
+
+---
+
+# Mit Content Slots
+
+In einem ListItem kann zusätzlicher `<Content />` (Top und Bottom Content)
+platziert werden. Die Position wird über das `slot`-Property gesteuert.
+
+<LiveCodeEditor example="item-content" editorCollapsed />
+
+---
+
+# Mit ColumnLayout
+
+Dem ListItem können die
+[ColumnLayout](/components/structure/column-layout)-Properties `s`, `m` und `l`
+mitgegeben werden, um Seitenverhältnis und Umbruchverhalten von Header und
+Content zu steuern.
+
+<LiveCodeEditor example="item-column-layout" editorCollapsed stretch />
+
+Da für die Spalten auch `null` gesetzt werden kann, lässt sich nicht zwingend
+benötigter Content in kleineren Ansichten ausblenden. In diesem Fall werden auch
+die entsprechenden Content Slots nicht angezeigt.
+
+<LiveCodeEditor example="item-column-layout-hidden" editorCollapsed stretch />
+
+---
+
+# Mit ActionGroup
+
+Verwende eine ActionGroup innerhalb des `<Content />`, um
+[Buttons](/components/actions/button) im ListItem zu platzieren.
+
+<LiveCodeEditor example="item-action-group" editorCollapsed stretch />
