+++ apps/docs/src/content/components/list/list/examples/domainApi.ts
@@ -1,0 +1,1 @@
+../../../../../../../../dev/domainApi.ts
