+++ apps/docs/src/content/components/list/list/examples/filter.tsx
@@ -14,7 +14,7 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list/examples/domainApi";
 
 export default () => {
   const DomainList = typedList<Domain>();
@@ -36,12 +36,18 @@ export default () => {
         mode="some"
         name="Typ"
       />
+      <DomainList.Filter
+        property="tld"
+        mode="some"
+        name="TLD"
+        priority="secondary"
+      />
       <DomainList.Sorting
         property="hostname"
         name="Alphabetisch"
         direction="asc"
-        directionName="aufsteigend"
         defaultEnabled
+        directionName="aufsteigend"
       />
       <DomainList.Sorting
         property="hostname"
