+++ apps/docs/src/content/components/list/list/examples/loadingView.tsx
@@ -10,7 +10,7 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list/examples/domainApi";
 
 export default () => {
   const List = typedList<Domain>();
