+++ apps/docs/src/content/components/list/list/examples/namingDont.tsx
@@ -12,23 +12,44 @@ import {
 import {
   type Domain,
   domains,
-} from "@/content/components/structure/list/examples/domainApi";
+} from "@/content/components/list/list/examples/domainApi";
 
 export default () => {
   const DomainList = typedList<Domain>();
 
   return (
     <DomainList.List
-      batchSize={2}
+      batchSize={5}
       aria-label="Domains"
       getItemId={(domain) => domain.id}
-      hidePagination
     >
       <DomainList.StaticData data={domains} />
-      <DomainList.Sorting
-        property="domain"
-        name="Name"
-        defaultEnabled
+      <DomainList.Filter
+        property="type"
+        mode="some"
+        name="Typ"
+        matcher={(filterValue, propertyValue) =>
+          filterValue === "Type Domain"
+            ? propertyValue === "Domain"
+            : propertyValue === "Subdomain"
+        }
+        values={["Type Domain", "Type Subdomain"]}
+        defaultSelected={["Type Domain"]}
+      />
+      <DomainList.Filter
+        property="verified"
+        mode="some"
+        name="Verifizierung"
+        matcher={(filterValue, propertyValue) =>
+          filterValue === "Verifizierung Verifiziert"
+            ? propertyValue
+            : !propertyValue
+        }
+        defaultSelected={["Verifizierung Unverifiziert"]}
+        values={[
+          "Verifizierung Verifiziert",
+          "Verifizierung Unverifiziert",
+        ]}
       />
       <DomainList.Item
         textValue={(domain) => domain.domain}
