+++ apps/docs/src/content/components/structure/flex/index.mdx
@@ -58,8 +58,8 @@ Breakpoints korrekt.
 
 ## List
 
-Innerhalb einer [List](/components/structure/list) platzierst du mit `<Flex />`
-den Content eines ListItems gezielt – in diesem Beispiel den zweiten Text in der
+Innerhalb einer [List](/components/list/list) platzierst du mit `<Flex />` den
+Content eines ListItems gezielt – in diesem Beispiel den zweiten Text in der
 rechten unteren Ecke.
 
 <LiveCodeEditor example="listItem" editorCollapsed />
