+++ apps/docs/src/content/foundations/content-guidelines/informationskonzept/index.mdx
@@ -52,8 +52,8 @@ in der Nähe des betroffenen Inhalts platziert werden (s.
 Eine [AlertBadge](/components/status/alert-badge) dient dazu, den Status eines
 Elements schnell sichtbar zu machen, beispielsweise in
 [Headings](/components/content/heading) oder
-[List](/components/structure/list)-Einträgen. In vielen Fällen wird sie durch
-einen [Alert](/components/status/alert) ergänzt, etwa auf der
+[List](/components/list/list)-Einträgen. In vielen Fällen wird sie durch einen
+[Alert](/components/status/alert) ergänzt, etwa auf der
 [Detailseite](/patterns/patterns/detail-page), um zusätzliche Informationen
 bereitzustellen. Für Metadaten oder ergänzende Informationen ist hingegen die
 [Badge](/components/status/badge) besser geeignet (siehe
