+++ apps/docs/src/content/foundations/structure/layout/index.mdx
@@ -79,9 +79,9 @@ auf anderen Seiten verwendet wurde.
 
 Die Navigationsstruktur im Content-Bereich folgt ebenfalls einem klaren Muster.
 Im Content-Bereich können [Tabs](/components/structure/tabs) und
-[List](/components/structure/list) verwendet werden, um die Inhalte zu teilen
-oder in die Tiefe zu navigieren. Bei der Navigation in tiefere Ebenen, z. B.
-über eine List, bietet eine [Breadcrumb](/components/navigation/breadcrumb) eine
+[List](/components/list/list) verwendet werden, um die Inhalte zu teilen oder in
+die Tiefe zu navigieren. Bei der Navigation in tiefere Ebenen, z. B. über eine
+List, bietet eine [Breadcrumb](/components/navigation/breadcrumb) eine
 Orientierungshilfe.
 
 Der Content wird von einer [LayoutCard](/components/structure/layout-card)
