+++ apps/docs/src/content/get-started/installation/index.mdx
@@ -47,10 +47,9 @@ Dafür umschließt du deine Anwendung mit dem `<ComponentDefaultsProvider />`:
 ```
 
 Jede einzelne Verwendung bleibt stärker als der Standard: Setzt eine
-[List](/components/structure/list) das `disableInitialSuspenseBoundary`-Property
-an ihrer Datenquelle, gilt dort dieser Wert. Verschachtelte Provider gelten
-jeweils für ihren Teilbaum und übernehmen die Werte, die sie nicht selbst
-nennen.
+[List](/components/list/list) das `disableInitialSuspenseBoundary`-Property an
+ihrer Datenquelle, gilt dort dieser Wert. Verschachtelte Provider gelten jeweils
+für ihren Teilbaum und übernehmen die Werte, die sie nicht selbst nennen.
 
 Diese Defaults stehen zur Verfügung:
