+++ apps/docs/src/content/patterns/patterns/anlegeprozess/index.mdx
@@ -93,8 +93,8 @@ lassen sich nach dem Anlegen jederzeit auf der Detailseite anpassen.
 
 Ein leerer Zustand bedeutet: Es wurden noch keine Elemente angelegt. In diesem
 Fall ersetzt die [IllustratedMessage](/components/content/illustrated-message)
-die [List](/components/structure/list) und bietet einen großen CTA zum Anlegen
-des ersten Elements.
+die [List](/components/list/list) und bietet einen großen CTA zum Anlegen des
+ersten Elements.
 
 **Beispiel**: Sind noch keine Projekte angelegt, wird in der Projektübersicht
 eine [IllustratedMessage](/components/content/illustrated-message) angezeigt.
@@ -106,7 +106,7 @@ eine [IllustratedMessage](/components/content/illustrated-message) angezeigt.
 />
 
 Sobald Inhalte vorhanden sind, rückt die Navigation über die
-[List](/components/structure/list) in den Vordergrund. Die
+[List](/components/list/list) in den Vordergrund. Die
 [ActionGroup](/components/actions/action-group) im Header der List übernimmt
 dann die Funktion, den Anlegeprozess für neue Elemente zu starten.
