+++ apps/docs/src/content/patterns/patterns/errorhandling/index.mdx
@@ -45,8 +45,7 @@ und bietet idealerweise eine Handlung an, zum Beispiel einen
 Für kleinere, abgegrenzte UI Elemente wie Tabellen- oder Listenzeilen sollte ein
 [AlertText](/components/status/alert-text) verwendet werden. So bleibt der
 Kontext erhalten und der Fehler wirkt sich nicht auf die ganze
-[Table](/components/structure/table) oder [List](/components/structure/list)
-aus.
+[Table](/components/structure/table) oder [List](/components/list/list) aus.
 
 <LiveCodeEditor example="alert-text" editorCollapsed bgColor="mstudio" />
