+++ apps/docs/src/content/patterns/patterns/list-tile-table/index.mdx
@@ -12,7 +12,7 @@ filtern und zu sortieren.
 
 # Zentrale Merkmale
 
-Die zwei Components [List](/components/structure/list) und
+Die zwei Components [List](/components/list/list) und
 [Table](/components/structure/table) dienen der kompakten Auflistung von
 Elementen. In der Verwendung ist die List jedoch häufiger relevanter.
 
@@ -66,7 +66,7 @@ Informationen, wodurch die Darstellung eher technisch wirkt. Sie wird vor allem
 für Detailinformationen genutzt, nicht für größere visuell hervorgehobene
 Elemente. Die Tabellenansicht verwendet die Component
 [Table](/components/structure/table) als Bestandteil der
-[List](/components/structure/list).
+[List](/components/list/list).
 
 <LiveCodeEditor example="ticket-list" editorCollapsed bgColor="mstudio" />
 
@@ -74,8 +74,8 @@ Elemente. Die Tabellenansicht verwendet die Component
 
 # Position und Aufbau
 
-[Lists](/components/structure/list) und [Tables](/components/structure/table)
-werden vor allem für die Aufzählung von Elementen verwendet – häufig als
+[Lists](/components/list/list) und [Tables](/components/structure/table) werden
+vor allem für die Aufzählung von Elementen verwendet – häufig als
 Einstiegsseiten zu [Detailansichten](/patterns/patterns/detail-page).
 
 Wenn die List oder Table auch ohne sichtbare
@@ -122,7 +122,7 @@ zusammengefasst werden. Filter sind in der Regel UND-verknüpft.
 Elemente innerhalb der List. Sie ergänzt die Sortier- und Filterfunktionen, um
 die Navigation besonders bei umfangreichen Listen effizient zu gestalten.
 
-**Actions:** Die [List](/components/structure/list) kann eine
+**Actions:** Die [List](/components/list/list) kann eine
 [ActionGroup](/components/actions/action-group) mit verschiedenen
 [Buttons](/components/actions/button) enthalten, die sich auf den Inhalt der
 List beziehen. Insbesondere Anlegeprozesse (Pattern:
@@ -135,7 +135,7 @@ einen Button gestartet werden.
 
 # Erstes Anlegen
 
-Wenn eine [List](/components/structure/list) oder
+Wenn eine [List](/components/list/list) oder
 [Table](/components/structure/table) noch keine Elemente enthält, sollte dies
 durch eine [IllustratedMessage](/components/content/illustrated-message) oder
 einen kurzen Hinweistext (wenn wenig Platz vorhanden) sichtbar gemacht werden.
