+++ apps/docs/src/content/patterns/patterns/navigation/index.mdx
@@ -103,7 +103,7 @@ Breite nicht aus, klappen überzählige Punkte automatisch in ein
 
 ### Listen, Raster und Tabellen
 
-Die verschiedenen Darstellungen einer [List](/components/structure/list)
+Die verschiedenen Darstellungen einer [List](/components/list/list)
 strukturieren Inhalte, schaffen Übersicht über ähnliche Elemente und ermöglichen
 die Navigation zu detaillierteren Ansichten.
